Python prefers modules over static methods
